Updated QuestionTypes, Quiz, User; changed id type to long

+++ server/src/main/java/pl/agh/hackathon/User/User.java
@@ -13,14 +13,12 @@ public class User
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(nullable = false)
     private String name;
 
     @Column(nullable = false)
     private String password;
-
-
-
+    
 }
